Format code correctly
